Basic Monster class skeleton
We still need logic for attack() and heal().

Currently, attack does not take a parameter for a target unlike Hero's attack.

The assumption is that there will only be one Hero and Monsters will only attack the singular Hero, but that may prove limiting. For now, it's probably safe to assume this.

As far as heal() goes we will need a method to set the DungeonCharacter's health while not exceeding a certain limit. Currently, we don't declare a "default" HealthPoints level to compare against and only track the "current" HealthPoints.

Perhaps when we switch to the SQLite database we can reference the default values.