Hero subclass attack method skeletons
The attack method signatures will likely will need modifications to match the Hero class (or conversely changing the Hero class to match the subclasses) in order to effectively use inheritance.

Currently, since the Hero attack signature does not match the subclasses the Hero's attack method is NOT overridden. I expect this can lead to great confusion at some point in the future.