Implement Item class to get Item name/description from DB
In the future we may choose to also get various stats modifiers and other parameters explaining the use of the items.

For instance, an item with positive health_points such as a potion could increase the health points of the character using it, and one with negative could be considered an offensive item which could be "used" on another character such as a monster.

For the stat_multiplier field of the item Table we may be able to implement a system where a character holding an item gains a certain stat boost (for instance after collecting a pillar)